All buttons of Canvas box are now working
All buttons of canvasBox are now working: additional functionalities to be added.
Snap-to-grid button: Snap to grid button now snaps the control points to the grid by 1 step, more settings to be added later. for UI the button changes to red while the icon changes to white, their hover changes to white, while the icon changes to red
Fit-to-screen button:  resets the zoom scale and the offset of the grid on the canvas. UI remains same
Save-button: save button loops over the preset array and checks if the current curve already is saved or not, if it is saved, then it changes UI and tooltip to indicate it, otherwise it allows you to click it and save it (name currently set to new preset by default; naming conditions to be decided later)