Save working properly, Unsave have problmes unsaving
Save button is working, UI is great. Unsave is having problem with the yes button because of the setTitle state variable. is not being reset, if it is reset to new preset, then the unsave yes button does not work, and if not reset, then the newly saved preset is named same a sthe previously saved preset.