Saving and Unsaving is fully functional
Saving and Unsaving have updated UI and are fully functional. Used a new function called areBezierValuesEqual to compare the values of the preset in order to save and unsave, and also used in various other places within the code for better comparasion.

Also updated the title changes of the animation cards. Curve card now dynamically changes title based on whether the values of the curve matches the value of a preset, preset also dynamically changes title based on selected preset.