manual value change functional
Users can now manually change the value of the control points to relocate them by double clicking a control point, which opens a prompt that allows users to write a x and a y value. X clamps to 0, and 1, respectively to the cosest value if it is out of bounds.